scripts/requirements-export.txt
# Optional: for export_elevenlabs_chapter1.py --chapter-yaml (PyYAML).
PyYAML>=6.0.1
